requirements.txt
yfinance
pandas

google-generativeai
requests
python-dotenv
